+++ Session Solution/C#Advanced/Assignment2/Q2/IRectangle.cs
@@ -4,7 +4,7 @@ namespace Assignment2.Q2;
 
 public interface IRectangle : IShape
 {
-    public decimal Length {get; set;}
-    public decimal Width {get; set;}
+    public double Length {get; set;}
+    public double Width {get; set;}
 
 }
